+++ src/components/select/multi-select/multi-select.spec.js
@@ -103,6 +103,38 @@ describe("MultiSelect", () => {
   });
 
   it.each([
+    "auto",
+    "auto-start",
+    "auto-end",
+    "top",
+    "top-start",
+    "top-end",
+    "bottom",
+    "bottom-start",
+    "bottom-end",
+    "right",
+    "right-start",
+    "right-end",
+    "left",
+    "left-start",
+    "left-end",
+  ])("the listPlacement prop should be passed", (listPlacement) => {
+    const wrapper = renderSelect({ listPlacement });
+
+    wrapper.find(Textbox).find('[type="dropdown"]').first().simulate("click");
+    expect(wrapper.find(SelectList).prop("listPlacement")).toBe(listPlacement);
+  });
+
+  it("the flipEnabled prop should be passed", () => {
+    const wrapper = renderSelect({ flipEnabled: false });
+
+    wrapper.find(Textbox).find('[type="dropdown"]').first().simulate("click");
+    expect(wrapper.find(SelectList).prop("flipEnabled")).toBe(false);
+    wrapper.setProps({ flipEnabled: true });
+    expect(wrapper.find(SelectList).prop("flipEnabled")).toBe(true);
+  });
+
+  it.each([
     ["small", "32px"],
     ["medium", "40px"],
     ["large", "48px"],