+++ src/components/select/select-list/select-list-container.style.js
@@ -8,9 +8,18 @@ const StyledSelectListContainer = styled.div`
   ${({ placement }) => placement === "top-start" && "bottom: 0"};
   min-width: 100%;
   max-width: 870px;
-  transition: height 0.15s ease-out;
   height: ${({ height }) => height};
   overflow: hidden;
+  animation: fadeIn 250ms ease-out;
+
+  @keyframes fadeIn {
+    0% {
+      opacity: 0;
+    }
+    100% {
+      opacity: 1;
+    }
+  }
 `;
 
 StyledSelectListContainer.defaultProps = {