+++ src/components/select/simple-select/simple-select.d.ts
@@ -37,6 +37,25 @@ export interface SimpleSelectProps
   value?: string | Record<string, unknown>;
   /** Overrides the default tooltip position */
   tooltipPosition?: "top" | "bottom" | "left" | "right";
+  /** Placement of the select list in relation to the input element */
+  listPlacement?:
+    | "auto"
+    | "auto-start"
+    | "auto-end"
+    | "top"
+    | "top-start"
+    | "top-end"
+    | "bottom"
+    | "bottom-start"
+    | "bottom-end"
+    | "right"
+    | "right-start"
+    | "right-end"
+    | "left"
+    | "left-start"
+    | "left-end";
+  /** Use the opposite list placement if the set placement does not fit */
+  flipEnabled?: bool;
 }
 
 declare function SimpleSelect(